Added SongBank and Song classes
The Song class just stores the value for the title, value, and
difficulty of the song to make them easier to access.

The SongBank class creates 5 instances of Song and makes a list of all
the songs. It also houses a getter method for this list, and a method
that takes in the title of a song and returns the song object that
corresponds to that title.