Note animation progress, made note and life images
Ive begun animating notes. The Song class is now in use for each song
value and working perfectly. The display notes function must be in its
own thread so that I can sleep the thread with no effect on the Main
thread. The timing is working very well. There is one issue with this
commit- when the note is first displayed, the AnimationJFrame compresses
into a tiny box. If you expand the window you can see that the
displayNotes function is working perfectly as is the timing, but this
has to be fixed obviously.

The first note images have been made- one for sharps and flats, one for
standard keys.
The first life image was made- I will probably end up splitting this
into two separate images so that only two hearts are required on the
screen- with each failure, half the heart disappears.