Many changes outlined below;
In AnimitaionJFrame, contants were added for the window width and window
height as these properties were used by multiple other classes. To fix,
the window sizing issue, I added a setsize method call just under
pack();. I must call pack so that all the components can be seen, but
this also resizes the display which is not what I want. Therefore, this
addition fixed the issue. For all notes to be displayed, I also added a
layout- Springlayout to the JFrame.

In Board, I had to be able to change the image of the note based on
whether or not it is a black or white key. I built two methods- the
first is my own array contains method, the second is a method that
returns the image path based on the INITIAL_X value by using the
contains method.

In Main, the animation of notes is complete. It takes in the chosen
song and converts the characters in the songs value string into initial
x values which are then passed onto the animateNote method. This is done
in it's own thread as if the character is a '.', the thread must sleep
for 1000 ms. This is how timing/tempo is achieved in this game.
In addition, the JFugue player now plays it's notes in their own thread.
A method was made to make calling the thread shorter in the switch
statement. This enables notes to be played closer together for a much
more realistic feel, however two notes still cannot be played at once.