Note colour change, bugs
So in this commit, now the note colour changes when you succesffully hit
the right key at the right time- or so I hoped. It seems to only work
well on the first hit. Extremely buggy.
In addition, I added a variable- livesRemaining which does keep good
track of remaining lives.