Piano improvements and max life modification
Previously, there were two hashmaps of data regarding notes on the
JFrame. I used these to figure out which note it being pressed as a key
is pressed (if any). It was terrible and extremely buggy.

Now, there is only one hashmap- a Pair of integers(the note position)
that is the key for the note object itself. While this did shorten the
code in many areas, in main I had to write another method, identifyNote,
to iterate through the notes in the new hashmap and return a note(if
any) that has the matching x value AND is in the y value range(in other
words, the note.canHit is true!).

The result of this major change is that the piano key recognition is
much less buggy and can actually recognise multiple consecutive notes.
There are still kinks that I am not sure whether it is user error as I
play the game or down to the code.

The smaller change is that the lifeRemaining changes at the very start
based on the difficulty of the song. If EASY, the user gets 6 lives to
start. If MEDIUM, 4. If HARD, 2.