Aesthetics update
Multiple changes made to set up screens quicker; most changes are for
making the game look better. In this commit, the selection screen
recieved the most changes. It now matches the other screens in colour
and has an image for the button. The colours match and it has its own
font. Very nice looking compared to before.
All screens now display the proper icon in the upper left.
All images added in this commit have also been added to the README.md.